+++ src/components/Home/Video.js
@@ -1,25 +1,17 @@
 import { faPlayCircle } from '@fortawesome/free-regular-svg-icons';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
-import Tippy from '@tippyjs/react';
 import { useState } from 'react';
 import { Button, Modal, ModalBody, ModalHeader } from 'reactstrap';
 import styled from 'styled-components';
 
 const PlayButton = styled(Button)`
-    font-size: 0.9rem !important;
     color: inherit !important;
-    text-decoration: none !important;
+    text-decoration: underline;
     transition: opacity 0.2s !important;
-    text-align: center !important;
+    font-size: inherit !important;
 
     &:hover {
         opacity: 0.7;
-    }
-
-    svg {
-        font-size: 65px !important;
-        line-height: normal;
-        margin-left: 3px;
     }
 `;
 
@@ -27,13 +19,9 @@ const Video = () => {
     const [isOpen, setIsOpen] = useState(false);
 
     return (
-        <div>
-            <PlayButton color="link" className="p-0 m-0" onClick={() => setIsOpen(true)}>
-                <Tippy content="Play introduction video" offset={[0, 50]}>
-                    <span>
-                        <Icon icon={faPlayCircle} />
-                    </span>
-                </Tippy>
+        <>
+            <PlayButton color="link" className="p-0 m-0 d-inline" onClick={() => setIsOpen(true)}>
+                <Icon icon={faPlayCircle} /> Play video
             </PlayButton>
 
             <Modal isOpen={isOpen} toggle={() => setIsOpen(v => !v)} size="lg">
@@ -52,7 +40,7 @@ const Video = () => {
                     </div>
                 </ModalBody>
             </Modal>
-        </div>
+        </>
     );
 };