+++ src/components/ViewPaper/EditDialog/hooks/useEditPaper.js
@@ -134,15 +134,21 @@ const useEditPaper = () => {
             [PREDICATES.URL]: 'url'
         };
 
-        for (const { predicate: property, object } of paperStatements) {
+        for (const { predicate: property, object, id: statementId } of paperStatements) {
             if (property.id in propertyIdToKey) {
-                data[propertyIdToKey[property.id]] = object;
+                data[propertyIdToKey[property.id]] = {
+                    ...object,
+                    statementId: property.id === PREDICATES.HAS_RESEARCH_FIELD || property.id === PREDICATES.HAS_VENUE ? statementId : undefined
+                };
             }
             if (property.id === PREDICATES.HAS_AUTHOR) {
-                data.authors.push(object);
+                data.authors.push({
+                    ...object,
+                    statementId
+                });
             }
         }
-
+        data.authors.reverse();
         return data;
     };